Changing comments and naming conventions
