+++ .gitignore
@@ -1,3 +1,5 @@
+.idea
+.idea/**/*
 /vendor
 composer.phar
 composer.lock
