+++ src/Spinegar/Sugar7Wrapper/Clients/Sugar7.php
@@ -1,0 +1,298 @@
+<?php namespace Spinegar\Sugar7Wrapper\Clients;
+
+use Guzzle\Common\Event;
+use Guzzle\Http\Client;
+use Guzzle\Http\Query;
+
+/**
+ * SugarCRM 7 Rest Client
+ *
+ * @package   SugarCRM 7 Rest Wrapper
+ * @category  Libraries
+ * @author  Sean Pinegar
+ * @license MIT License
+ * @link   https://github.com/spinegar/sugarcrm7-api-wrapper-class
+ */
+class Sugar7 implements ClientInterface
+{
+
+    /**
+     * Variable: $url
+     * Description:  A Sugar Instance.
+     */
+    private $url;
+
+    /**
+     * Variable: $username
+     * Description:  A SugarCRM User.
+     */
+    private $username;
+
+    /**
+     * Variable: $password
+     * Description:  The password for the $username SugarCRM account
+     */
+    private $password;
+
+    /**
+     * The client id used for sugar when identifying this connection
+     * @var string
+     */
+    private $clientId = 'sugar';
+
+    /**
+     * The client secret key for authenticating connections with sugar
+     * @var string
+     */
+    private $clientSecret = '';
+
+    /**
+     * Variable: $platform
+     * Description:  A Sugar Instance.
+     */
+    private $platform = 'api';
+
+    /**
+     * Variable: $token
+     * Description:  OAuth 2.0 token
+     */
+    protected $token;
+
+    /**
+     * Variable: $client
+     * Description:  Guzzle Client
+     */
+    protected $client;
+
+    function __construct()
+    {
+        $this->client = new Client();
+    }
+
+    public function getNewAuthToken()
+    {
+        $request = $this->client->post('oauth2/token', null, array(
+            'grant_type' => 'password',
+            'client_id' => $this->client_id,
+            'client_secret' => $this->client_secret,
+            'username' => $this->username,
+            'password' => $this->password,
+            'platform' => $this->platform,
+        ));
+
+        $result = $request->send()->json();
+        return $result['access_token'];
+    }
+
+    public function connect()
+    {
+        $token = $this->getNewAuthToken();
+
+        if (!$token) {
+            return false;
+        }
+
+        self::setToken($token);
+        $eventDispatcher = $this->client->getEventDispatcher();
+        $eventDispatcher->addListener('request.before_send', array($this, 'beforeSendRequest'));
+        $eventDispatcher->addListener('request.error', array($this, 'refreshToken'));
+
+        return true;
+    }
+
+    public function check()
+    {
+        if (!$this->token)
+            return false;
+
+        return true;
+    }
+
+    public function setClientOption($key, $value)
+    {
+        if (!$key || $value)
+            return false;
+
+        $this->client->setDefaultOption($key, $value);
+
+        return true;
+    }
+
+    public function setUrl($url)
+    {
+        if (!$url) {
+            return false;
+        }
+
+        $this->url = $url;
+        $this->client->setBaseUrl($this->url);
+
+        return true;
+    }
+
+    public function getUrl()
+    {
+        return $this->url;
+    }
+    public function setUsername($value)
+    {
+        $this->username = $value;
+    }
+
+    public function setPassword($value)
+    {
+        $this->password = $value;
+    }
+
+    public function setClientId($value)
+    {
+        $this->clientId = $value;
+    }
+
+    public function setClientSecret($value)
+    {
+        $this->clientSecret = $value;
+    }
+
+    public function setPlatform($value)
+    {
+        if (!$value)
+            return false;
+
+        $this->platform = $value;
+
+        return true;
+    }
+
+    public function getPlatform()
+    {
+        return $this->platform;
+    }
+
+    public function setToken($value)
+    {
+        if (!$value)
+            return false;
+
+        $this->token = $value;
+
+        return true;
+    }
+
+    public function get($endpoint, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->get($endpoint);
+
+        $query = $request->getQuery();
+
+        foreach ($parameters as $key => $value) {
+            $query->add($key, $value);
+        }
+
+        $response = $request->send()->json();
+
+        if (!$response)
+            return false;
+
+        return $response;
+    }
+
+    public function getFile($endpoint, $destinationFile, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->get($endpoint);
+
+        $query = $request->getQuery();
+
+        foreach ($parameters as $key => $value) {
+            $query->add($key, $value);
+        }
+
+        $request->setResponseBody($destinationFile);
+
+        $response = $request->send();
+
+        if (!$response)
+            return false;
+
+        return $response;
+    }
+
+    public function postFile($endpoint, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->post($endpoint, array(), $parameters);
+        $request->setHeader('Content-Type', 'multipart/form-data');
+        $result = $request->send();
+
+        if (!$result)
+            return false;
+
+        return $result;
+    }
+
+    public function post($endpoint, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->post($endpoint, null, json_encode($parameters));
+        $response = $request->send()->json();
+
+        if (!$response)
+            return false;
+
+        return $response;
+    }
+
+    public function put($endpoint, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->put($endpoint, null, json_encode($parameters));
+        $response = $request->send()->json();
+
+        if (!$response)
+            return false;
+
+        return $response;
+    }
+
+    public function delete($endpoint, $parameters = array())
+    {
+        if (!self::check())
+            self::connect();
+
+        $request = $this->client->delete($endpoint);
+        $response = $request->send()->json();
+
+
+        if (!$response)
+            return false;
+
+        return $response;
+    }
+
+    public function refreshToken(Event $event)
+    {
+        if ($event['response']->getStatusCode() === 401) {
+            $this->setToken($this->getNewAuthToken());
+
+            $event['response'] = $event['request']->send();
+            $event->stopPropagation();
+        }
+    }
+
+    public function beforeSendRequest(Event $event)
+    {
+        $event['request']->setHeader('OAuth-Token', $this->token);
+    }
+}
